+++ NEWS.md
@@ -25,6 +25,8 @@
 * Fix `IdfViewer$snapshot()` for PNG output. A new parameter `webshot` has been
   added in `IdfViewer$snapshot()` with default value being `FALSE`. (#509).
 * Now `IdfSchedule$... <- val` works (#512).
+* Fix the error of invalid datetime in `Epw$data()` when data of Feb 28 is from
+  a leap year while the EPW did not contain any leap year data (#552).
 
 # eplusr 0.15.1